fixed ambiguous conversion for static_cast from 'size_t' (aka 'unsigned long long') to 'hip_fp8_e4m3' (aka '__hip_fp8_e4m3')

+++ projects/hipdnn/test_sdk/tests/utilities/TestCpuFpReferenceConvolution.cpp
@@ -1114,7 +1114,7 @@ TEST(TestCpuFpReferenceConvolutionFp8, ConvolutionBwdDataBasic)
     outputTensor.setHostValue(4.0_fp8, 0, 0, 1, 1);
 
     // Weight values: simple 3x3 kernel
-    for(size_t i = 0; i < 9; ++i)
+    for(int i = 0; i < 9; ++i)
     {
         weightTensor.memory().hostData()[i] = static_cast<hip_fp8_e4m3>(i + 1);
     }
